Refactor collectionHasUnstavedArticleEditsWarningSelector
